Rework scaling intersection edge detection based on resolution.
Depth doesn't really work, since the effect of depth on how many pixels we have for
an object depends heavily on the camera FOV.  Figure out the pixel density instead
and scale based on that.  The intended effect is to scale up the intersection threshold
when we don't actually have enough pixels in an area to reliably distinguish intersections
from gaps due to lack of resolution, and this should work regardless of whether it's
caused by an object being far away or the image simply being low-resolution.