Make masks optionally per-layer.
If you have lots of objects, and you want to use a mask with just one of them,
having masks always global can slow things down a lot applying it to a bunch
of object IDs you don't need.  Make --layer-mask apply to the most recent
--layer.

If a --layer-mask comes before any layers, it'll be applied to all layers like
before.

To apply a mask only to the default layer, you'll need to explicitly create the
layer (object 0) rather than letting it be added automatically.